move Filter trait into serde module
